third_party/WebKit/LayoutTests/http/tests/websocket/cookie-ws-to-ws.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script src="resources/get-request-header.js"></script>
<script>
description('Test receiving cookies in a WebSocket handshake which were set by another WebSocket handshake.');

window.jsTestIsAsync = true;

// Normalize a cookie string
function normalizeCookie(cookie) {
    // Split the cookie string, sort it and then put it back together.
    return cookie.split('; ').sort().join('; ');
}

var cookie;
var URL_SET_COOKIE = 'ws://127.0.0.1:8880/set-cookie';
var ws_set_cookie = new WebSocket(URL_SET_COOKIE);

ws_set_cookie.onopen = function () {
    ws_set_cookie.close();
};
ws_set_cookie.onclose = function (e) {
    connectAndGetRequestHeader('cookie').then(function(value) {
        cookie = normalizeCookie(value);
        shouldBeEqualToString('cookie', 'WK-WebSocket-test-domain-pass=1; WK-WebSocket-test-path-pass=1; WK-WebSocket-test=1');
        clear();
    }, finishAsFailed);
};

function clear() {
    var ws = new WebSocket(URL_SET_COOKIE + '?clear=1');
    ws.onopen = function () {
        ws.close();
    };
    ws.onclose = finishJSTest;
}
setTimeout(finishJSTest, 1000);
</script>
